+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Update documentation
 - Minor formatting fixes
+- Remove `Moar.Term.or_default/2` since it is the same as `Moar.Term.presence/2`.
 
 ## 0.1.0